refactor: P2 server split 7/N — extract learner profile route domain
Move all /api/learner/* routes (profile GET/PATCH, goals POST/PATCH/DELETE,
knowledge PATCH, misconceptions PATCH) out of server.ts into
server/routes/learner.ts as handleLearnerRoutes. The domain-private helpers
clamp01 and normalizePreferences move with it; everything else the routes
touch (loadProfile/saveProfile, learner types, randomUUID) was already
imported from memory/learner and is now imported directly by the routes
file. server.ts: 2758 → 2612 lines.

Smoke coverage: goal create → patch → delete round-trip (404 after delete).
